requirements.txt
streamlit
openai
python-dotenv
faiss-cpu
numpy